+++ src/Sharpliner/GitHubActions/Runner.cs
@@ -1,4 +1,4 @@
-namespace Sharpliner.GHActions
+namespace Sharpliner.GitHubActions
 {
     public record Runner
     {
